Move exception related code to own subdir
