+++ include/argent.h
@@ -27,8 +27,8 @@
 #include "./primitives.h"
 #include "./registry.h"
 #include "./log.h"
-#include "./erno.h"
-#include "./exception.h"
+#include "../src/ex/erno.h"
+#include "../src/ex//exception.h"
 #include "./hash.h"
 #include "./string.h"
 #include "./memblock.h"
