+++ src/ex/handler.c
@@ -21,7 +21,7 @@
  ******************************************************************************/
 
 
-#include "../include/argent.h"
+#include "../../include/argent.h"
 
 #include <regex.h>
